attempting a V2 New, Errorf, Wrap, WrapPrefix return error, not *Error Move all code to error.go, all tests to error_test.go Remove error_1_13* and join_unwrap_*.  We will assume go 1.20+
internal helper method wrap() for Wrap() and WrapPrefix(), needed to adjust
"skip" depth.

Changing the semantics of Is().  From the doc for errors.Is:
> An Is method should only shallowly compare err and the target and not
> call Unwrap on either.
The previous Is examined (*Error).Err, which is effectively inlining
Unwrap().  Other standard go errors do not do this.  Is(err1, err2) should
strictly determine if err2 is within the lineage of err1, and not examine if
err1 is in the lineage of err2
Updated tests accordingly