initial work for V2
remove all organization for GO version dependencies and backwards compatibility;
we depend on Go 1.20
all functions return error, not *Error
refactor Wrap() and WrapPrefix() with a private wrap().  We need wrap() to
return *Error so that Wrap/WrapPrefix can access type-specific fields
Remove custom Is() code.  The code was effectively unwrapping the *Error, which
the Go Is() documentation states we should not do: https://pkg.go.dev/errors#Is
"An Is method should only shallowly compare err and the target and not call
Unwrap on either." . By accessing (*Error.Err) we are unwrapping.  In the old
code Is() and As() were inconsistent (Is might return true but then As failed to
find an error)

Update README for TODOs.

The module remains github.com/go-errors/errors
This can be used by other modules by using the replace directive in go.mod